+++ changelog.md
@@ -5,7 +5,7 @@
 ### Version 6.0.0
 
 * Requires Gradle 6.6+.
-* `quasar-utils`: Upgrade to use Quasar 0.8.2_r3 OSGi bundles.
+* `quasar-utils`: Upgrade to use Quasar 0.8.5_r3 OSGi bundles.
 * `quasar-utils`: Support configurable `@Suspendable` annotation.
 * `cordapp`: Generate CPK format CorDapps.